# docs/swagger.yaml
basePath: /
definitions:
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Attendance:
    properties:
      created_at:
        type: string
      date:
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      session:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
      session_id:
        type: string
      status:
        description: '''present'', ''absent'', ''late'', ''excused'''
        type: string
      updated_at:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.AttendanceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Attendance'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CategoryListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CategoryResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class:
    properties:
      capacity:
        type: integer
      category:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category'
      category_id:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      price:
        description: 'Financial Data: Always use int64 for money, price, balance,
          or amounts'
        type: integer
      tenant_id:
        type: string
      type:
        description: '''private'' or ''group'''
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      day_of_week:
        description: 1 = Senin, 7 = Minggu (ISO 8601)
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: standard HH:MM:SS for time column
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      location:
        type: string
      start_time:
        description: standard HH:MM:SS for time column
        type: string
      tutor_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      end_time:
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
      schedule_id:
        type: string
      session_date:
        type: string
      start_time:
        type: string
      status:
        description: '''scheduled'', ''rescheduled'', ''cancelled'', ''completed'''
        type: string
      tutor_id:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateAttendanceRequest:
    properties:
      date:
        type: string
      enrollment_id:
        type: string
      schedule_id:
        type: string
      status:
        enum:
        - present
        - absent
        - late
        - excused
        type: string
    required:
    - date
    - enrollment_id
    - schedule_id
    - status
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateCategoryRequest:
    properties:
      description:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - name
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      joined_at:
        type: string
      status:
        description: '''pending'', ''active'', ''completed'', ''dropped'''
        type: string
      student:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
      student_id:
        type: string
      tenant_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse:
    properties:
      data: {}
      message:
        example: An error occurred
        type: string
      status:
        example: error
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      full_name:
        type: string
      id:
        type: string
      parent_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateAttendanceRequest:
    properties:
      status:
        enum:
        - present
        - absent
        - late
        - excused
        type: string
    required:
    - status
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Academic Management Service for KelolaKelas Platform
  title: KelolaKelas Academic Service API
  version: "1.0"
paths:
  /api/v1/attendance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.AttendanceListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List attendance
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      parameters:
      - description: Attendance payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateAttendanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Attendance'
              type: object
      security:
      - BearerAuth: []
      summary: Create attendance
      tags:
      - Attendance
  /api/v1/attendance/{id}:
    get:
      parameters:
      - description: Attendance UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Attendance'
              type: object
      security:
      - BearerAuth: []
      summary: Get attendance
      tags:
      - Attendance
    patch:
      consumes:
      - application/json
      parameters:
      - description: Attendance UUID
        in: path
        name: id
        required: true
        type: string
      - description: Attendance payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Attendance'
              type: object
      security:
      - BearerAuth: []
      summary: Update attendance
      tags:
      - Attendance
  /api/v1/categories:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Search category name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CategoryListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tenant categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new subject/course category for a tenant
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Create category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create academic category
      tags:
      - Categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete a tenant-owned category when it has no active classes
      parameters:
      - description: Tenant ID dalam format UUID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Category ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete academic category
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"